Create new QR using react-query done
