+++ src/hooks/useAPI/useQuoteRequests.jsx
@@ -1,5 +1,6 @@
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import {
+  createNewQuoteRequest,
   deleteQuoteRequestById,
   fetchAllQuoteRequests,
   updateQuoteRequestById
@@ -13,6 +14,34 @@ export const useFetchAllQuoteRequests = () => {
   });
 
   return { data, isLoading, isError };
+};
+
+// Custom react-query hook to create a new quote request
+export const useCreateNewQuoteRequest = (toast) => {
+  const queryClient = useQueryClient();
+  return useMutation((newQuoteRequestObject) => createNewQuoteRequest(newQuoteRequestObject), {
+    onError: (error) => {
+      toast({
+        position: 'top',
+        title: `Error Occured Creating New QR`,
+        description: `Error: ${error.message}`,
+        status: 'error',
+        duration: 5000,
+        isClosable: true
+      });
+    },
+    onSuccess: (data) => {
+      queryClient.invalidateQueries({ queryKey: ['quoteRequests'] });
+      toast({
+        position: 'top',
+        title: `Quote Request created!`,
+        description: `We've created a new quote request for ${data.firstName} ${data.lastName} with email ${data.email} 🚀`,
+        status: 'success',
+        duration: 5000,
+        isClosable: true
+      });
+    }
+  });
 };
 
 // Custom react-query hook for deleting a qr by id
